refactor para funcionar como paquete java y en vscode
